Changed: Move `com.termux.app.utils.CrashUtils` to `com.termux.shared.termux.crash.TermuxCrashUtils` so that plugins trigger plugin notifications too
Calls to `notifyAppCrashFromCrashLogFile()` will now be synchronized as well.
